Extract approved domains tests into a separate file
